refactor out things for the sample.py code
